// frontend/src/app/api/proxies/issueTrackerAPI.schemas.ts
export interface Member {
  userId: string;
  email: string;
  /** @nullable */
  name: string | null;
  /** @nullable */
  imageUrl: string | null;
  role: string;
  permissions: string[];
  createdAt: string;
}

export interface MemberList {
  members: Member[];
}

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  admin: 'admin',
  member: 'member',
  viewer: 'viewer',
} as const;